Reorganization of newevent, code cleanup
