+++ src/pages/Home/StyleLanding.js
@@ -201,6 +201,15 @@ export const EventItems = styled.div`
 export const EventItem = styled.div`
   display: flex;
   width: 90%;
+  &:hover {
+    cursor: pointer;
+    text-decoration: underline;
+  }
+`;
+
+export const FakeEventItem = styled.div`
+  display: flex;
+  width: 90%;
 `;
 
 export const EventDescription = styled.div`
